utworzenie repository, service,exception, controller oraz testy

+++ src/main/java/com/example/s27553_bank/Model/Account.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class Account {
     private String name;
     private String surname;
-    private Integer accountsaldo;
+    private Double accountsaldo;
     private String pesel;
     private Currency currency;
 }
